+++ kinto_http/batch.py
@@ -19,10 +19,16 @@ class BatchSession(object):
         self._results = []
 
     def request(self, method, endpoint, data=None, permissions=None,
-                headers=None):
+                payload=None, headers=None):
         # Store all the requests in a dict, to be read later when .send()
         # is called.
-        self.requests.append((method, endpoint, data, permissions, headers))
+        payload = payload or {}
+        if data is not None:
+            payload['data'] = data
+        if permissions is not None:
+            payload['permissions'] = permissions
+
+        self.requests.append((method, endpoint, payload, headers))
         # This is the signature of the session request.
         return defaultdict(dict), defaultdict(dict)
 
@@ -32,17 +38,13 @@ class BatchSession(object):
 
     def _build_requests(self):
         requests = []
-        for (method, url, data, permissions, headers) in self.requests:
+        for (method, url, payload, headers) in self.requests:
             # Strip the prefix in batch requests.
             request = {
                 'method': method.upper(),
                 'path': url.replace('v1/', '')}
 
-            request['body'] = {}
-            if data is not None:
-                request['body']['data'] = data
-            if permissions is not None:
-                request['body']['permissions'] = permissions
+            request['body'] = payload
             if headers is not None:
                 request['headers'] = headers
             requests.append(request)